+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/internal/KtorNetworkTransport.kt
@@ -36,9 +36,10 @@ import io.ktor.utils.io.errors.IOException
 import io.realm.kotlin.internal.interop.sync.NetworkTransport
 import io.realm.kotlin.internal.interop.sync.Response
 import io.realm.kotlin.internal.interop.sync.ResponseCallback
+import io.realm.kotlin.internal.util.CoroutineDispatcherFactory
+import io.realm.kotlin.internal.util.DispatcherHolder
 import io.realm.kotlin.mongodb.AppConfiguration.Companion.DEFAULT_AUTHORIZATION_HEADER_NAME
 import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.async
 import kotlin.collections.set
@@ -49,10 +50,11 @@ public class KtorNetworkTransport(
     // FIXME Rework timeout to take a Duration instead
     //  https://github.com/realm/realm-kotlin/issues/408
     timeoutMs: Long,
-    private val dispatcher: CoroutineDispatcher,
+    dispatcherFactory: CoroutineDispatcherFactory,
     logger: Logger? = null,
 ) : NetworkTransport {
 
+    private val dispatcherHolder: DispatcherHolder = dispatcherFactory.create()
     private val clientCache: HttpClientCache = HttpClientCache(timeoutMs, logger)
 
     @Suppress("ComplexMethod", "TooGenericExceptionCaught")
@@ -64,7 +66,7 @@ public class KtorNetworkTransport(
         callback: ResponseCallback,
     ) {
         val client = clientCache.getClient()
-        CoroutineScope(dispatcher).async {
+        CoroutineScope(dispatcherHolder.dispatcher).async {
             val response = try {
                 val requestBuilderBlock: HttpRequestBuilder.() -> Unit = {
                     headers {
@@ -124,6 +126,7 @@ public class KtorNetworkTransport(
 
     override fun close() {
         clientCache.close()
+        dispatcherHolder.close()
     }
 
     private suspend fun processHttpResponse(response: HttpResponse): Response {